+++ AgentPlatform-BackEnd/server/app/modules/runs/schemas.py
@@ -19,7 +19,7 @@ class RunToolCall(WireModel):
     args: dict[str, Any]
     result: Any | None = None
     error: str | None = None
-    duration_ms: int
+    duration_ms: float
     status: CallStatus
